requirements.txt
langchain==0.0.131
openai==0.27.2
pinecone-client==2.2.1
unstructured==0.5.9
